[diffusion] MiniMax-H3: move the AdaLN cache into its own module
Pure relocation out of the 2.6k-line minimax_h3.py ahead of the
tiered-cache work; no behavior change.
